pyproject.toml
[project]
name = "email-validator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dnspython>=2.7.0",
]
